% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{align.local.all}
\alias{align.local.all}
\title{All local alignments}
\usage{
align.local.all(clust1, clust2, mat, threshold, cores = 2, dismat = NULL)
}
\arguments{
\item{clust1}{a list of clusters}

\item{clust2}{a list of clusters}

\item{mat}{a similarity matrix}

\item{threshold}{a threshold}

\item{cores}{number of cores}

\item{dismat}{a disimilarity matrix to use in the local aligments}
}
\value{
a list of alignments
}
\description{
Compute a list of local alignments between \code{clust1} and
\code{clust2}. The function compute all the local alignments of all pairs
of clusters whose centers have a similarity greather than \code{ll}
}

