% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{align.local.plot}
\alias{align.local.plot}
\title{Local alignment plot
plot a local alignment and all the local alignments that
intersect with him}
\usage{
align.local.plot(localAligns, global, p1, p2, net1, net2, ...)
}
\arguments{
\item{localAligns}{a list of local alignments}

\item{global}{an alignment}

\item{p1}{the center of cluster1}

\item{p2}{the center of cluster2}

\item{net1}{the first network}

\item{net2}{the second network}

\item{...}{further arguments to be passed to igraph plotting}
}
\description{
Local alignment plot
plot a local alignment and all the local alignments that
intersect with him
}

