% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{align.local}
\alias{align.local}
\title{Local alignment}
\usage{
align.local(net1, net2, p1, p2, compute.ec = FALSE, mat = NULL)
}
\arguments{
\item{net1}{network}

\item{net2}{network}

\item{p1}{center of net1}

\item{p2}{center of net2}

\item{compute.ec}{compute ecscore (TRUE/FALSE)}

\item{mat}{a non dissimilarity matrix}
}
\value{
alignment
}
\description{
Compute the local alignment between the networks \code{net1} and
\code{net2} with centers \code{p1} and \code{p2}
}

