% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{align.plot}
\alias{align.plot}
\title{Plot the alignment}
\usage{
align.plot(net1, net2, global, k1 = 1, k2 = 1, edge.curved = 0.5, ...)
}
\arguments{
\item{net1}{an igraph object}

\item{net2}{an igrpah object}

\item{global}{an alignment}

\item{k1}{the width of the new edges of the alignment}

\item{k2}{the width of the old edges}

\item{edge.curved}{Specifies whether to draw curved
edges, or not. This can be a logical or a numeric vector or scalar.}

\item{...}{further arguments to be passed to igraph plotting}
}
\description{
Plot the alignment
}

