% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{compute.matrix}
\alias{compute.matrix}
\title{Compute Matrix}
\usage{
compute.matrix(net1, net2 = NULL, type = "Distance", mode = "Similarity",
  database = NULL, database2 = NULL, byComp = TRUE, normalized = TRUE)
}
\arguments{
\item{net1}{an igraph object}

\item{net2}{an igraph object for modes \code{BLAST} and \code{FC} if Net2
is not NULL, computes the matrix between nodes in Net1 and Net2.}

\item{type}{what matrice do you want to compute (BLAST, DSD, Distance,
FC, Degree)}

\item{mode}{for type \code{BLAST} : pident or bitscore.
For \code{DSD} and \code{Distance} distance, similarity or similarity by
components.
For \code{FC} the list of ontologies}

\item{database}{for \code{BLAST} and \code{FC} the database which
proteins belongs.}

\item{database2}{for \code{BLAST} and \code{FC} the database which
proteins belongs.}

\item{byComp}{for \code{DSD} and \code{Distance} if the similarity or
distance must be normalized by components}

\item{normalized}{if the matrix will be normalized}
}
\value{
The matrix
}
\description{
Compute several matrices
}

