% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dme-dme}
\alias{dme-dme}
\alias{dme.dme}
\title{Similarity matrix between proteins of dme}
\format{a \code{matrix} instance}
\description{
A similarity matrix between the proteins of dme. The similarity was calculated
as an average between the blast bitscore normalized of the sequence of the proteins and the
similarity of the degrees. The second part was calculated using the function
\code{compute.matrix}
}
\author{
Adria Alcala, 07/12/2015
}

