% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dme-sce}
\alias{dme-sce}
\alias{dme.sce}
\title{Similarity matrix between proteins of dme and sce}
\format{a \code{matrix} instance}
\description{
A similarity matrix between the proteins of dme. The similarity is
the blast bitscore normalized of the sequence of the proteins.
}
\author{
Adria Alcala, 07/12/2015
}

