% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dme}
\alias{dme}
\title{Protein protein interaction network of dme}
\format{a \code{igraph} instance}
\description{
The biigest connected component of the protein protein interaction
network of drosophila melanogaster. Extracted from IsoBase
}
\author{
Adria Alcala, 07/12/2015
}

