% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{EC.score.all}
\alias{EC.score.all}
\title{Edges score}
\usage{
EC.score.all(localAligns, net1, net2)
}
\arguments{
\item{localAligns, }{a list of local alignments}

\item{net1}{an igraph object}

\item{net2}{an igraph object}
}
\value{
edgesScore  a table with 3 columns, the first and second ones
represents the local alignment and the third the score of alignment
}
\description{
Compute the EC.score of the list of local alignments
}

