% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{extract.clusters}
\alias{extract.clusters}
\title{Extract clusters}
\usage{
extract.clusters(Net, ClustMat)
}
\arguments{
\item{Net}{an igraph object}

\item{ClustMat}{a cluster matrix (output of \code{cluster.matrix})}
}
\value{
Clusters in igraph format
}
\description{
Compute the subnetworks of \code{Net} from a cluster matrix
}

