% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{FC.score}
\alias{FC.score}
\title{Functional Coherence Score}
\usage{
FC.score(alin, gos)
}
\arguments{
\item{alin}{alignment, with the format of \code{alin.local} or
\code{alin.global}}

\item{gos}{a list of ontologies}
}
\value{
FC score
}
\description{
Given an alignment \eqn{g:V_1 \rightarrow V_2}, and an ontologies \eqn{gos},
then the functional coherence score is defined as:
\deqn{\frac{1}{|V_1|} \sum_{v \in V_1} \frac{|gos(v) \cap gos(g(v))|}
{|gos(v)\cup gos(g(v))|}}
}

