% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{go}
\alias{go}
\title{A list of go}
\format{a \code{list} instance}
\source{
IsoBase
}
\description{
A list of the gene ontology of the proteins in dme and sce protein
protein interaction network
}
\author{
Adria Alcala, 07/12/2015
}

