% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{read.matrix}
\alias{read.matrix}
\title{Read Matrix}
\usage{
read.matrix(fileName, mode = "table", sym = "TRUE", def = 0)
}
\arguments{
\item{fileName}{directory of the matrix or data.frame (for mode col3)}

\item{mode}{table, col3, RData}

\item{sym}{True if the matrix must be symmetric.In modes table and RData,
if the matrix is not symmetric the value of M[i,j] will be (M[i,j]+M[j,i])/2.
In mode in 3col the value of M[i,j] will be the last that appears in the file}

\item{def}{the default value in mode 3col}
}
\value{
matrix
}
\description{
Funció per a llegir una matriu des de un fitxer. Accepta diversos formats:
- table. A la linea i del fitxer hi ha el valor de tota la fila i de la matriu.
- col3. A cada linea del fitxer hi ha les columnes fila - columna - valor
- RData. El fitxer és un objecte de R
}
\details{
També accepta els parametres sym, per si es vol convertir la matriu a una matriu simètrica i
def el valor per defecte que tendrà la matriu en les posicions que en el format col3 no
estan definides.
Aquesta funció també es pot utilitzar per passar una matriu que estigui a un data.frame en
el format col3 a matriu.
}

