% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{read.network}
\alias{read.network}
\title{Read Network}
\usage{
read.network(fileName, mode = "tab", db = "uniprot", cols = "all",
  sep = " ", int.type = TRUE)
}
\arguments{
\item{fileName}{directory of network or a data.frame}

\item{mode}{tab : a file from DIP or BIOGRID of type tab or mitab;
edges: a list of edges with attributes (not necessary)}

\item{db}{database to select proteins of mode tab (uniprot,DIP,refseq,BIOGRID, gene)}

\item{cols}{which columns must be selected with mode edges, you can use "all" to
select all columns}

\item{sep}{separator between columns in edge mode}

\item{int.type}{boolean to decide if must be select the interaction type in tab mode.}
}
\description{
Funcio per a llegir un graf desde un fitxer o un data.frame. Accepta dos tipus de formats
- tab : Un fitxer que prove de DIP o BIOGRIDP en format tab o mitab
- edges: Un conjunt d'arestes, les quals poden tenir atributs. Els atributs estaran a partir
de la tercera columna. Es a dir, el format ha de ser node1 - node2 - atribut1 - atribut2 - ...
}
\details{
Si es llegeix amb el format tab, s'ha d'indicar de quina base de dades s'han de seleccionar
les proteines. En el cas del format tab, es pot indicar si es dessitja conserva com a atribut
de la interaccio el tipus d'interaccio.
En el format edges, es pot passar una llista de columnes, per a triar quines amb quin ordre
s'han de seleccionar per a la construccio del graf.
}

