% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{search.aligns}
\alias{search.aligns}
\title{Search alignments
Search alignments in als from p1 to p2}
\usage{
search.aligns(als, p1, p2)
}
\arguments{
\item{als}{a list of local alignments}

\item{p1}{a protein}

\item{p2}{a protein}
}
\value{
list of alignments that includes p1 and p2
}
\description{
Search alignments
Search alignments in als from p1 to p2
}

