% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{select.aligns}
\alias{select.aligns}
\title{Select alignments}
\usage{
select.aligns(localAligns, scores)
}
\arguments{
\item{localAligns}{a list of local alignments}

\item{scores}{a table of scores}
}
\value{
selectAligns a list of local alignments
}
\description{
Select a list of aligments that recover all the proteins, based on the
list of scores
}

