% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{sim.score.all}
\alias{sim.score.all}
\title{Similarity Score}
\usage{
sim.score.all(localAligns, sim)
}
\arguments{
\item{localAligns}{a list of local alignments}

\item{sim}{a similarity matrix}
}
\value{
simScore a table with 3 columns, the first and second ones
represents the local alignment and the third the score of alignment
}
\description{
Compute the similarity score of the list of local alignments. Given an
alignment \eqn{g:V_1\rightarrow V_2}, and a similarity matrix \eqn{sim}, the
similarity score of the alignment \eqn{g} is defined as:
\deqn{\frac{1}{|V_1|} \sum_{v \in V_1} sim[v,g(v)]}
}

