% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alin.R
\name{align.global}
\alias{align.global}
\title{Global Alignment of two Protein Interaction Network}
\usage{
align.global(localAligns, Sim)
}
\arguments{
\item{localAligns}{a list of local alignments}

\item{Sim}{a similarity matrix}
}
\value{
Global a list of two global alignments, the first one is the
solution of the hypergraph matching problem, and the second one is the
extension of this alignment
}
\description{
Return a global alignment, from the list of local alignments and the table
of scores. The function first calculate a list of alignments with
\code{selectAligns}, then found a solution of the hypergraph mathching
problem. And finally extend this alignment to a global alignment
}

