% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{cluster.network}
\alias{cluster.network}
\title{Cluster Network}
\usage{
cluster.network(sigma, lambda = 0, k = dim(sigma)[1])
}
\arguments{
\item{sigma}{a similarity matrix}

\item{lambda}{the similarity threshold}

\item{k}{the size threshold}
}
\value{
The cluster matrix, where M[i,j] is TRUE if node i belongs to the
cluster of node j and FALSE otherwise
}
\description{
Calculate a network clusters from a similarity matrix \code{sigma}. The output is a matrix,in which
the position (i, j) represents if the node i belongs to cluster j.
}
\details{
The cluster of a node, are the nodes in the network with a similarity greather than
\code{lambda}. In case
Els cluster d'un node i son els nodes que tenen una similaritat amb i major que lambda.
If there is more than \code{k} nodes with greater similarity than \code{lambda} the
\code{k} nodes more similars to i are selected.
}

