% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{compute.matrix.Distance}
\alias{compute.matrix.Distance}
\title{DSD}
\usage{
compute.matrix.Distance(net, mode = "Similarity", byComp = TRUE,
  normalized = TRUE)
}
\description{
DSD
}
\keyword{Distance}
\keyword{internal}

