% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{extract.clusters}
\alias{extract.clusters}
\title{Extract clusters}
\usage{
extract.clusters(Net, ClustMat)
}
\arguments{
\item{Net}{an igraph object}

\item{ClustMat}{a cluster matrix (output of \code{cluster.matrix})}
}
\value{
Clusters in igraph format
}
\description{
Donada una matriu que representa els clusters d'una xarxa, calcula les subxarxes que
representen aquests clusters. Es a dir, donat un node i, la subxarxa que representa el
cluster de i, es el subgraf de la xarxa que conte els nodes que pertanyen al cluster de i.
}
\details{
Si no es dona cap matriu es calcula la matriu amb la funcio cluster.network i els valors
lambda = 0.2 i k = 20.
Compute the subnetworks of \code{Net} from a cluster matrix
}

