% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{read.matrix}
\alias{read.matrix}
\title{Read Matrix}
\usage{
read.matrix(fileName, mode = "table", sym = "TRUE", def = 0)
}
\arguments{
\item{fileName}{directory of the matrix or data.frame (for mode col3)}

\item{mode}{table, col3, RData}

\item{sym}{True if the matrix must be symmetric.In modes table and RData,
if the matrix is not symmetric the value of M[i,j] will be (M[i,j]+M[j,i])/2.
In mode in 3col the value of M[i,j] will be the last that appears in the file}

\item{def}{the default value in mode 3col}
}
\value{
matrix
}
\description{
Read a matrix in several formats
}

