% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{read.network}
\alias{read.network}
\title{Read Network}
\usage{
read.network(fileName, mode = "tab", db = "uniprot", cols = "all",
  sep = " ", int.type = TRUE)
}
\arguments{
\item{fileName}{directory of network or a data.frame}

\item{mode}{tab : a file from DIP or BIOGRID of type tab or mitab;
edges: a list of edges with attributes (not necessary)}

\item{db}{database to select proteins of mode tab (uniprot,DIP,refseq,BIOGRID, gene)}

\item{cols}{which columns must be selected with mode edges, you can use "all" to
select all columns}

\item{sep}{separator between columns in edge mode}

\item{int.type}{boolean to decide if must be select the interaction type in tab mode.}
}
\description{
Function to read a graph from a file or a \code{date.frame}. Accepts two kinds of formats:
- \code{tab} : A file that comes from DIP or BioGRID in tab or mitab format.
- \code{edges}: Each row of the file or \code{data.frame} is an edge. The edges can have
attributes.
}
\details{
In \code{tab} format, you can indicate wheter you wish to preserves interaction type as
attribute.
In \code{edges} format, you can choose wich columns are the attributes of the edge.
}

